requirements.txt
# Core testing framework
robotframework==7.3.2

# SSH library and its dependencies
robotframework-sshlibrary==3.8.0
paramiko==3.5.1
scp==0.15.0
bcrypt==4.3.0
cryptography==45.0.4
PyNaCl==1.5.0

# For reading YAML config files
PyYAML==6.0.2